requirements.txt
requests==2.27.1
PyYAML==6.0
